Sync default controls
